Improve exit code handling
When using aws-profile to run commands which return exit codes > 128 the
exit code is shifted to return just the first 8 bits (See the shells and
scripts section of https://en.wikipedia.org/wiki/Exit_status#Shell_and_scripts).
This is due to the use of os.WEXITSTATUS

Switching to use sys.exit causes the actual return code of the called
script to be returned. This was brought to light when attempting to use
the awscli and the return code was 255, but the aws-profile returned
zero. This resulted in incorrect control flow in the wrapping script.