Introduce cleaner selection of AWS profiles
As mention in https://github.com/jrstarke/aws-profile/pull/5 botocore already handles picking and selecting AWS profile names from
AWS_PROFILE and AWS_DEFAULT_PROFILE so removing them and introducing a
named argument as a way of overriding environment variables seems to
make the usage and testability cleaner.

This is a breaking change so anyone who was using positional argument
for the profile will have issues when upgrading.